pyproject.toml
[project]
name = "grpc_python_hello_world"
version = "0.1.0"
description = "Hello, world example for gRPC in Python"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio-tools>=1.71.0",
]
